Add user management command and ability to get users by ID
- Make GetUser take an interface identifier that can be either a
  string (email) or int (ID).
- Add GetInput commands helper to encapsulate reading user input.
- Add GetPassword commands helper to encapsulate reading password
  input.
- Add user command with list, get, add, change, and delete sub-
  commands.

GetInput and GetPassword will likely be replaced in the future once
the shell package has a built-in API for querying for additional
input outside of the main loop.

With this additional complexity it may be time to begin separating
some of the commands, this one in particular, into different files.